# [Feature] Adopt mmeval's SSIM and SNR in mmediting