--- a/mmedit/evaluation/metrics/snr.py
+++ b/mmedit/evaluation/metrics/snr.py
@@ -1,30 +1,21 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import Optional
+from typing import Optional, Sequence
 
-import numpy as np
+from mmeval import SNR as SNR_MMEVAL
 
 from mmedit.registry import METRICS
-from .base_sample_wise_metric import BaseSampleWiseMetric
-from .metrics_utils import img_transform
+from .metrics_utils import obtain_data
 
 
 @METRICS.register_module()
-class SNR(BaseSampleWiseMetric):
-    """Signal-to-Noise Ratio.
+class SNR(SNR_MMEVAL):
+    """Signal-to-Noise Ratio. A wrapper of :class:`mmeval.SNR`.
 
     Ref: https://en.wikipedia.org/wiki/Signal-to-noise_ratio
 
     Args:
-
         gt_key (str): Key of ground-truth. Default: 'gt_img'
         pred_key (str): Key of prediction. Default: 'pred_img'
-        collect_device (str): Device name used for collecting results from
-            different ranks during distributed training. Must be 'cpu' or
-            'gpu'. Defaults to 'cpu'.
-        prefix (str, optional): The prefix that will be added in the metric
-            names to disambiguate homonymous metrics of different evaluators.
-            If prefix is not provided in the argument, self.default_prefix
-            will be used instead. Default: None
         crop_border (int): Cropped pixels in each edges of an image. These
             pixels are not involved in the SNR calculation. Default: 0.
         input_order (str): Whether the input order is 'HWC' or 'CHW'.
@@ -33,6 +24,11 @@ class SNR(BaseSampleWiseMetric):
             If None, the images are not altered. When computing for 'Y',
             the images are assumed to be in BGR order. Options are 'Y' and
             None. Default: None.
+        channel_order (str): The channel order of image. Default: 'rgb'.
+        prefix (str, optional): The prefix that will be added in the metric
+            names to disambiguate homonymous metrics of different evaluators.
+            If prefix is not provided in the argument, self.default_prefix
+            will be used instead. Default: None
 
     Metrics:
         - SNR (float): Signal-to-Noise Ratio
@@ -43,88 +39,60 @@ class SNR(BaseSampleWiseMetric):
     def __init__(self,
                  gt_key: str = 'gt_img',
                  pred_key: str = 'pred_img',
-                 collect_device: str = 'cpu',
-                 prefix: Optional[str] = None,
-                 crop_border=0,
                  input_order='CHW',
-                 convert_to=None) -> None:
-        super().__init__(
-            gt_key=gt_key,
-            pred_key=pred_key,
-            mask_key=None,
-            collect_device=collect_device,
-            prefix=prefix)
+                 crop_border=0,
+                 convert_to=None,
+                 channel_order: str = 'rgb',
+                 prefix: Optional[str] = None,
+                 scaling: float = 1,
+                 **kwargs) -> None:
+        super().__init__(input_order, convert_to, crop_border, channel_order,
+                         **kwargs)
 
-        self.crop_border = crop_border
-        self.input_order = input_order
-        self.convert_to = convert_to
+        self.gt_key = gt_key
+        self.pred_key = pred_key
+        self.prefix = prefix
+        self.scaling = scaling
 
-    def process_image(self, gt, pred, mask):
-        """Process an image.
+    def process(self, data_batch: Sequence[dict],
+                data_samples: Sequence[dict]) -> None:
+        """Process one batch of data and predictions.
 
         Args:
-            gt (Torch | np.ndarray): GT image.
-            pred (Torch | np.ndarray): Pred image.
-            mask (Torch | np.ndarray): Mask of evaluation.
-        Returns:
-            np.ndarray: SNR result.
+            data_batch (Sequence[dict]): A batch of data
+                from the dataloader.
+            predictions (Sequence[dict]): A batch of outputs from
+                the model.
         """
 
-        return snr(
-            gt=gt,
-            pred=pred,
-            crop_border=self.crop_border,
-            input_order=self.input_order,
-            convert_to=self.convert_to,
-            channel_order=self.channel_order)
+        for data in data_samples:
+            prediction = data['output']
 
+            channel_order = 'rgb'
+            metainfo = data
+            if 'gt_channel_order' in metainfo:
+                channel_order = metainfo['gt_channel_order']
+            elif 'img_channel_order' in metainfo:
+                channel_order = metainfo['img_channel_order']
 
-def snr(gt,
-        pred,
-        crop_border=0,
-        input_order='HWC',
-        convert_to=None,
-        channel_order='rgb'):
-    """Calculate PSNR (Peak Signal-to-Noise Ratio).
+            gt = obtain_data(data, self.gt_key)
+            pred = obtain_data(prediction, self.pred_key)
 
-    Ref: https://en.wikipedia.org/wiki/Peak_signal-to-noise_ratio
+            gt = [sample for sample in gt]
+            pred = [sample for sample in pred]
 
-    Args:
-        gt (ndarray): Images with range [0, 255].
-        pred (ndarray): Images with range [0, 255].
-        crop_border (int): Cropped pixels in each edges of an image. These
-            pixels are not involved in the PSNR calculation. Default: 0.
-        input_order (str): Whether the input order is 'HWC' or 'CHW'.
-            Default: 'HWC'.
-        convert_to (str): Whether to convert the images to other color models.
-            If None, the images are not altered. When computing for 'Y',
-            the images are assumed to be in BGR order. Options are 'Y' and
-            None. Default: None.
-        channel_order (str): The channel order of image. Default: 'rgb'.
+            self.add(gt, pred, channel_order)
 
-    Returns:
-        float: SNR result.
-    """
+    def evaluate(self, *args, **kwargs):
+        """Returns metric results and print pretty table of metrics per class.
 
-    assert gt.shape == pred.shape, (
-        f'Image shapes are different: {gt.shape}, {pred.shape}.')
+        This method would be invoked by ``mmengine.Evaluator``.
+        """
+        metric_results = self.compute(*args, **kwargs)
+        self.reset()
 
-    gt = img_transform(
-        gt,
-        crop_border=crop_border,
-        input_order=input_order,
-        convert_to=convert_to,
-        channel_order=channel_order)
-    pred = img_transform(
-        pred,
-        crop_border=crop_border,
-        input_order=input_order,
-        convert_to=convert_to,
-        channel_order=channel_order)
-
-    signal = ((gt)**2).mean()
-    noise = ((gt - pred)**2).mean()
-
-    result = 10. * np.log10(signal / noise)
-
-    return result
+        key_template = f'{self.prefix}/{{}}' if self.prefix else '{}'
+        return {
+            key_template.format(k): v * self.scaling
+            for k, v in metric_results.items()
+        }