--- a/mmedit/evaluation/metrics/ssim.py
+++ b/mmedit/evaluation/metrics/ssim.py
@@ -1,18 +1,16 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import Optional
+from typing import Optional, Sequence
 
-import cv2
-import numpy as np
+from mmeval import SSIM as SSIM_MMEVAL
 
 from mmedit.registry import METRICS
-from mmedit.utils import to_numpy
-from .base_sample_wise_metric import BaseSampleWiseMetric
-from .metrics_utils import img_transform
+from .metrics_utils import obtain_data
 
 
 @METRICS.register_module()
-class SSIM(BaseSampleWiseMetric):
-    """Calculate SSIM (structural similarity).
+class SSIM(SSIM_MMEVAL):
+    """Calculate SSIM (structural similarity). A wrapper of
+    :class:`mmeval.SSIM`.
 
     Ref:
     Image quality assessment: From error visibility to structural similarity
@@ -24,16 +22,8 @@ class SSIM(BaseSampleWiseMetric):
     averaged.
 
     Args:
-
         gt_key (str): Key of ground-truth. Default: 'gt_img'
         pred_key (str): Key of prediction. Default: 'pred_img'
-        collect_device (str): Device name used for collecting results from
-            different ranks during distributed training. Must be 'cpu' or
-            'gpu'. Defaults to 'cpu'.
-        prefix (str, optional): The prefix that will be added in the metric
-            names to disambiguate homonymous metrics of different evaluators.
-            If prefix is not provided in the argument, self.default_prefix
-            will be used instead. Default: None
         crop_border (int): Cropped pixels in each edges of an image. These
             pixels are not involved in the PSNR calculation. Default: 0.
         input_order (str): Whether the input order is 'HWC' or 'CHW'.
@@ -42,141 +32,77 @@ class SSIM(BaseSampleWiseMetric):
             If None, the images are not altered. When computing for 'Y',
             the images are assumed to be in BGR order. Options are 'Y' and
             None. Default: None.
+        channel_order (str): The channel order of image. Default: 'rgb'.
+        prefix (str, optional): The prefix that will be added in the metric
+            names to disambiguate homonymous metrics of different evaluators.
+            If prefix is not provided in the argument, self.default_prefix
+            will be used instead. Default: None
+        scaling (float, optional): Scaling factor for final metric.
+            E.g. scaling=100 means the final metric will be amplified by 100
+            for output. Default: 1
+        **kwargs: Keyword parameters passed to :class:`BaseMetric`.
 
     Metrics:
         - SSIM (float): Structural similarity
     """
 
-    metric = 'SSIM'
-
     def __init__(self,
                  gt_key: str = 'gt_img',
                  pred_key: str = 'pred_img',
-                 collect_device: str = 'cpu',
+                 input_order: str = 'CHW',
+                 crop_border: int = 0,
+                 convert_to: Optional[str] = None,
+                 channel_order: str = 'rgb',
                  prefix: Optional[str] = None,
-                 crop_border=0,
-                 input_order='CHW',
-                 convert_to=None) -> None:
-        super().__init__(
-            gt_key=gt_key,
-            pred_key=pred_key,
-            mask_key=None,
-            collect_device=collect_device,
-            prefix=prefix)
+                 scaling: float = 1,
+                 **kwargs) -> None:
+        super().__init__(input_order, crop_border, convert_to, channel_order,
+                         **kwargs)
 
-        self.crop_border = crop_border
-        self.input_order = input_order
-        self.convert_to = convert_to
+        self.gt_key = gt_key
+        self.pred_key = pred_key
+        self.prefix = prefix
+        self.scaling = scaling
 
-    def process_image(self, gt, pred, mask):
-        """Process an image.
+    def process(self, data_batch: Sequence[dict],
+                data_samples: Sequence[dict]) -> None:
+        """Process one batch of data and predictions.
 
         Args:
-            gt (Torch | np.ndarray): GT image.
-            pred (Torch | np.ndarray): Pred image.
-            mask (Torch | np.ndarray): Mask of evaluation.
-        Returns:
-            np.ndarray: SSIM result.
+            data_batch (Sequence[dict]): A batch of data
+                from the dataloader.
+            predictions (Sequence[dict]): A batch of outputs from
+                the model.
         """
 
-        return ssim(
-            img1=gt,
-            img2=pred,
-            crop_border=self.crop_border,
-            input_order=self.input_order,
-            convert_to=self.convert_to,
-            channel_order=self.channel_order)
+        for data in data_samples:
+            prediction = data['output']
 
+            channel_order = 'rgb'
+            metainfo = data
+            if 'gt_channel_order' in metainfo:
+                channel_order = metainfo['gt_channel_order']
+            elif 'img_channel_order' in metainfo:
+                channel_order = metainfo['img_channel_order']
 
-def _ssim(img1, img2):
-    """Calculate SSIM (structural similarity) for one channel images.
+            gt = obtain_data(data, self.gt_key)
+            pred = obtain_data(prediction, self.pred_key)
 
-    It is called by func:`ssim`.
+            gt = [sample.numpy() for sample in gt]
+            pred = [sample.numpy() for sample in pred]
 
-    Args:
-        img1, img2 (np.ndarray): Images with range [0, 255] with order 'HWC'.
+            self.add(gt, pred, channel_order)
 
-    Returns:
-        float: SSIM result.
-    """
+    def evaluate(self, *args, **kwargs):
+        """Returns metric results and print pretty table of metrics per class.
 
-    C1 = (0.01 * 255)**2
-    C2 = (0.03 * 255)**2
+        This method would be invoked by ``mmengine.Evaluator``.
+        """
+        metric_results = self.compute(*args, **kwargs)
+        self.reset()
 
-    kernel = cv2.getGaussianKernel(11, 1.5)
-    window = np.outer(kernel, kernel.transpose())
-
-    mu1 = cv2.filter2D(img1, -1, window)[5:-5, 5:-5]
-    mu2 = cv2.filter2D(img2, -1, window)[5:-5, 5:-5]
-    mu1_sq = mu1**2
-    mu2_sq = mu2**2
-    mu1_mu2 = mu1 * mu2
-    sigma1_sq = cv2.filter2D(img1**2, -1, window)[5:-5, 5:-5] - mu1_sq
-    sigma2_sq = cv2.filter2D(img2**2, -1, window)[5:-5, 5:-5] - mu2_sq
-    sigma12 = cv2.filter2D(img1 * img2, -1, window)[5:-5, 5:-5] - mu1_mu2
-
-    ssim_map = ((2 * mu1_mu2 + C1) *
-                (2 * sigma12 + C2)) / ((mu1_sq + mu2_sq + C1) *
-                                       (sigma1_sq + sigma2_sq + C2))
-
-    return ssim_map.mean()
-
-
-def ssim(img1,
-         img2,
-         crop_border=0,
-         input_order='HWC',
-         convert_to=None,
-         channel_order='rgb'):
-    """Calculate SSIM (structural similarity).
-
-    Ref:
-    Image quality assessment: From error visibility to structural similarity
-
-    The results are the same as that of the official released MATLAB code in
-    https://ece.uwaterloo.ca/~z70wang/research/ssim/.
-
-    For three-channel images, SSIM is calculated for each channel and then
-    averaged.
-
-    Args:
-        img1 (ndarray): Images with range [0, 255].
-        img2 (ndarray): Images with range [0, 255].
-        crop_border (int): Cropped pixels in each edges of an image. These
-            pixels are not involved in the SSIM calculation. Default: 0.
-        input_order (str): Whether the input order is 'HWC' or 'CHW'.
-            Default: 'HWC'.
-        convert_to (str): Whether to convert the images to other color models.
-            If None, the images are not altered. When computing for 'Y',
-            the images are assumed to be in BGR order. Options are 'Y' and
-            None. Default: None.
-        channel_order (str): The channel order of image. Default: 'rgb'
-
-    Returns:
-        float: SSIM result.
-    """
-
-    assert img1.shape == img2.shape, (
-        f'Image shapes are different: {img1.shape}, {img2.shape}.')
-
-    img1 = img_transform(
-        img1,
-        crop_border=crop_border,
-        input_order=input_order,
-        convert_to=convert_to,
-        channel_order=channel_order)
-    img2 = img_transform(
-        img2,
-        crop_border=crop_border,
-        input_order=input_order,
-        convert_to=convert_to,
-        channel_order=channel_order)
-
-    img1 = to_numpy(img1)
-    img2 = to_numpy(img2)
-
-    ssims = []
-    for i in range(img1.shape[2]):
-        ssims.append(_ssim(img1[..., i], img2[..., i]))
-
-    return np.array(ssims).mean()
+        key_template = f'{self.prefix}/{{}}' if self.prefix else '{}'
+        return {
+            key_template.format(k): v * self.scaling
+            for k, v in metric_results.items()
+        }