Thanks for your contribution and we appreciate it a lot. The following instructions would make your pull request more healthy and more easily get feedback. If you do not understand some items, don't worry, just make the pull request and seek help from maintainers.

## Motivation

Support MMEval's SSIM and SNR metric in MMEditing-1.x.

## Comparsion with Default Implementation

* Config: `esrgan_psnr-x4c64b23g32_1xb16-1000k_div2k.py`
* Pretrain Model: [url](https://download.openmmlab.com/mmediting/restorers/esrgan/esrgan_x4c64b23g32_1x16_400k_div2k_20200508-f8ccaf3b.pth)
* Slightly modification of the config: Change `PSNR` metric to `SNR` metric to comparison the performance.
* Known difference between MMEval and MMEditing: MMEval use `np.float64` to calculate SNR metric, and MMEditing use `np.float32`.
* Unit test for MMEditing has not be revised.

### Single GPU

* MMEditing-1.x (9min 42s)
![single-mmediting](https://user-images.githubusercontent.com/28132635/196877286-e8db2e8e-0dff-40d7-abd4-c98a8fe3266a.png)

* MMEditing with MMEval (9min 50s)
![single-mmeval](https://user-images.githubusercontent.com/28132635/196882100-1913d9d5-6219-4c20-a04c-35c3f3f7d369.png)

### Multi GPUs (2 GPUs)

* MMEditing-1.x (4min 52s)
![multi-mmediting](https://user-images.githubusercontent.com/28132635/196877302-5da0495d-0ac2-4d55-853e-524e0926f8d4.png)

* MMEditing with MMEval (4min 52s)
![multi-mmeval](https://user-images.githubusercontent.com/28132635/196877511-3e243bae-7cb1-46ea-b4d6-9b0a40a41ed2.png)

## Who can help? @ them here!

## BC-breaking (Optional)

Does the modification introduce changes that break the backward-compatibility of the downstream repositories?
If so, please describe how it breaks the compatibility and how the downstream projects should modify their code to keep compatibility with this PR.

## Use cases (Optional)

If this PR introduces a new feature, it is better to list some use cases here, and update the documentation.

## Checklist

**Before PR**:

- [x] I have read and followed the workflow indicated in the [CONTRIBUTING.md](https://github.com/open-mmlab/mmediting/blob/master/.github/CONTRIBUTING.md) to create this PR.
- [x] Pre-commit or linting tools indicated in [CONTRIBUTING.md](https://github.com/open-mmlab/mmediting/blob/master/.github/CONTRIBUTING.md) are used to fix the potential lint issues.
- [x] Bug fixes are covered by unit tests, the case that causes the bug should be added in the unit tests.
- [x] New functionalities are covered by complete unit tests. If not, please add more unit test to ensure the correctness.
- [x] The documentation has been modified accordingly, including docstring or example tutorials.

**After PR**:

- [x] If the modification has potential influence on downstream or other related projects, this PR should be tested with some of those projects.
- [x] CLA has been signed and all committers have signed the CLA in this PR.